feat: add color mapping for canvas chart fields
- Extract color utility functions to separate util.ts file
- Add labelFirst prop to ColorInput component for inline display
- Integrate color mapping UI into field config inspector
- Enable per-value color customization for chart data series
